+++ src/utils/index.tsx
@@ -43,7 +43,7 @@ export interface InputProps {
    */
   disabled?: boolean;
   /**
-   * Specify a custom `id` for the `<input>`
+   * Specify a unique `id` for the `<input>`
    */
   id: string;
   /**