+++ CHANGELOG.md
@@ -1,3 +1,106 @@
+## [1.133.4](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.133.3...v1.133.4) (2025-04-04)
+
+
+### Bug Fixes
+
+* **icon:** L3-6336 handle icons that are not of 1x1 aspect ratio ([#580](https://github.com/PhillipsAuctionHouse/seldon/issues/580)) ([3085546](https://github.com/PhillipsAuctionHouse/seldon/commit/3085546665d011272af6928578c180911ebe96ca))
+
+## [1.133.3](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.133.2...v1.133.3) (2025-04-04)
+
+
+### Bug Fixes
+
+* **Icons:** L3-6336 fix fallback value for Icon color ([#579](https://github.com/PhillipsAuctionHouse/seldon/issues/579)) ([16dc0bc](https://github.com/PhillipsAuctionHouse/seldon/commit/16dc0bc2728d581c903c49919c1aa1e4a108b7e2))
+
+## [1.133.2](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.133.1...v1.133.2) (2025-04-02)
+
+
+### Bug Fixes
+
+* **lot-description:** L3-6287 Fix for Object Tile Title Text ([#576](https://github.com/PhillipsAuctionHouse/seldon/issues/576)) ([0bc8145](https://github.com/PhillipsAuctionHouse/seldon/commit/0bc814534dc5e6be27fcb37304b556f6d368738c))
+
+## [1.133.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.133.0...v1.133.1) (2025-04-01)
+
+
+### Bug Fixes
+
+* **Grid:** L3-6072  Revert Add support for xl breakpoints and colStart props ([#564](https://github.com/PhillipsAuctionHouse/seldon/issues/564))" ([#574](https://github.com/PhillipsAuctionHouse/seldon/issues/574)) ([e52b0a5](https://github.com/PhillipsAuctionHouse/seldon/commit/e52b0a5c0b069878cf477f85c39287dddb5e4f06))
+
+# [1.133.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.132.1...v1.133.0) (2025-04-01)
+
+
+### Bug Fixes
+
+* **IconGrid:** L3-3417 fix build issue ([#573](https://github.com/PhillipsAuctionHouse/seldon/issues/573)) ([8869d4c](https://github.com/PhillipsAuctionHouse/seldon/commit/8869d4c1af6363ddda46fe4b2fd2108bd4f61264))
+
+
+### Features
+
+* **Sale and Lot Icons:** L3-3417 icon updates ([#566](https://github.com/PhillipsAuctionHouse/seldon/issues/566)) ([4c88d48](https://github.com/PhillipsAuctionHouse/seldon/commit/4c88d48f24cc0865ee60385593d8257d84c2f621))
+
+## [1.132.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.132.0...v1.132.1) (2025-04-01)
+
+
+### Bug Fixes
+
+* **upgrade:** L3-6222 allow seldon to be used with phillips-public older React 18 version ([#567](https://github.com/PhillipsAuctionHouse/seldon/issues/567)) ([312d0b5](https://github.com/PhillipsAuctionHouse/seldon/commit/312d0b5dd2de424dfbe93fff98677ece64f088c0))
+
+# [1.132.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.131.0...v1.132.0) (2025-03-28)
+
+
+### Features
+
+* **textarea:** L3-5975 Add textarea component ([#551](https://github.com/PhillipsAuctionHouse/seldon/issues/551)) ([da4847d](https://github.com/PhillipsAuctionHouse/seldon/commit/da4847de26f3a9007b045d3388b057677f65117c))
+
+# [1.131.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.130.1...v1.131.0) (2025-03-26)
+
+
+### Features
+
+* **Grid:** L3-6072 Add support for xl breakpoints and colStart props ([#564](https://github.com/PhillipsAuctionHouse/seldon/issues/564)) ([a447d79](https://github.com/PhillipsAuctionHouse/seldon/commit/a447d79f47303770ca21ffc10a56fca88a946e8a))
+
+## [1.130.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.130.0...v1.130.1) (2025-03-26)
+
+
+### Bug Fixes
+
+* **Search:** L3-5572 Search Bar Removing Typed Search ([#563](https://github.com/PhillipsAuctionHouse/seldon/issues/563)) ([9d96b40](https://github.com/PhillipsAuctionHouse/seldon/commit/9d96b4053138727dad46f74fc5423f3bf8cd27d1))
+
+# [1.130.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.129.2...v1.130.0) (2025-03-21)
+
+
+### Features
+
+* **_button:** L3-6070 fixed height of button ([#562](https://github.com/PhillipsAuctionHouse/seldon/issues/562)) ([ede0cdf](https://github.com/PhillipsAuctionHouse/seldon/commit/ede0cdff814123458e2598c545b86a3834196275))
+
+## [1.129.2](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.129.1...v1.129.2) (2025-03-18)
+
+
+### Bug Fixes
+
+* **objecttile:** L3-5542 object tile style fixes ([#552](https://github.com/PhillipsAuctionHouse/seldon/issues/552)) ([6ee7e7a](https://github.com/PhillipsAuctionHouse/seldon/commit/6ee7e7a7dce037a66eabaa4ca6ffe446736598b3))
+
+## [1.129.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.129.0...v1.129.1) (2025-03-18)
+
+
+### Bug Fixes
+
+* **bidsnapshot:** L3-6073 bid snapshot style fix ([#554](https://github.com/PhillipsAuctionHouse/seldon/issues/554)) ([821cbe7](https://github.com/PhillipsAuctionHouse/seldon/commit/821cbe7b9164556fa99c7fe5a95e52925350a302))
+
+# [1.129.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.128.0...v1.129.0) (2025-03-17)
+
+
+### Features
+
+* **ContentPeek:** L3-5721 fixed component's logic ([#553](https://github.com/PhillipsAuctionHouse/seldon/issues/553)) ([62a701f](https://github.com/PhillipsAuctionHouse/seldon/commit/62a701fe634d966a5a389297bf701015422649e1))
+
+# [1.128.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.127.0...v1.128.0) (2025-03-17)
+
+
+### Features
+
+* **drawer:** L3-5957 Update drawer max width ([#548](https://github.com/PhillipsAuctionHouse/seldon/issues/548)) ([667e38c](https://github.com/PhillipsAuctionHouse/seldon/commit/667e38c467b40fc845d5c7e1d31c4d10a408b809))
+
 # [1.127.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.126.1...v1.127.0) (2025-03-13)